+++ trunk/brkga/graph.h
@@ -11,12 +11,12 @@ class Graph {
     public:
         explicit Graph(int nVertex);
         explicit Graph(const Graph& g);
-         ~Graph();
+        ~Graph();
         void AddEdge(int i, int j);
         void Print();
         void DeleteEdge(int i, int j);
-		    int GetCost(int i, int j) const;
-	private:
+        float GetCost(int i, int j) const;
+    private:
         int **adjacencyMatrix_;
         int nVertex_;
 };
